Clarify distinction from BSD's err.h
libbsd is the name of the library that provides BSD functions on Linux
systems, but it doesn't make the most sense to refer to the BSD
functions themselves by this name.

A link to the source was added not just for the sake of providing
context, but also to prevent doxygen from automatically linkiing "err.h"
to the err.h in this project. This causes confusion in this context.